Made the jcontroller from joint_p_ctrl.py into a submodule.
